cleanup: include whitespace in template
We need this for the case where the string within the conditional is
rendered. If the whitespace is removed, then the two package names are
essentially concatanated, which is not the behaviour we want.

So let us live with the additional whitespace in the output file till we
find a better mechanism to handle this.